3 fix cepat dari audit metodologi (B1, B4, C1)
B1: reseed torch.manual_seed(exp_cfg["seed"]) tepat sebelum training DeepMF (baik
model penuh maupun tiap fold OOF) di run_baseline.py/run_baseline_absa.py --
sebelumnya cuma di-seed sekali di awal pipeline, jadi bobot awal DeepMF bisa
berbeda antar run walau seed nominal sama (state RNG tergantung riwayat operasi
sebelumnya seperti preprocessing/ABSA). Sekarang konsisten dengan
tune_deepmf_oof_val.py yang sudah reseed di titik yang sama.

B4: coordinate search di tune_deepmf_oof_val.py sekarang bandingkan pemenang tiap
stage terhadap current_best_rmse yang dibawa dari stage sebelumnya (CLI baru
--base-rmse), bukan cuma sesama kandidat baru dalam stage yang sama. Sebelumnya
bug ini menyebabkan stage_adamw_epochs melaporkan epochs=5 sebagai "pemenang"
padahal epochs=20 (basis) sebenarnya masih lebih baik di val.

C1: InteractionDataset beri dtype=np.int64 eksplisit ke array negatif kosong
(negative_ratio=0) -- sebelumnya np.array([]) default float64 memaksa concatenate
dengan pos_users/pos_items (int64) jadi float64 diam-diam, sumber
DeprecationWarning berulang di semua run.

Referensi: reports/methodology_audit_2026-07-26.md

+++ src/baseline/deepmf.py
@@ -108,8 +108,16 @@ class InteractionDataset(Dataset):
 
         neg_labels = np.zeros(len(neg_users), dtype=np.float32)
 
-        self.users = np.concatenate([pos_users, np.array(neg_users)])
-        self.items = np.concatenate([pos_items, np.array(neg_items)])
+        # dtype=np.int64 EKSPLISIT (Temuan C1 audit metodologi): kalau
+        # negative_ratio=0 (dipakai di semua config saat ini, lihat
+        # docstring modul), neg_users/neg_items kosong -- np.array([]) TANPA
+        # dtype eksplisit default float64, memaksa np.concatenate() dgn
+        # pos_users/pos_items (int64) jadi float64 scr diam-diam. Tidak
+        # mengubah hasil (dikonversi balik ke torch.long di __getitem__),
+        # tapi memicu DeprecationWarning berulang & rapuh thd perubahan
+        # perilaku numpy/pytorch ke depan.
+        self.users = np.concatenate([pos_users, np.array(neg_users, dtype=np.int64)])
+        self.items = np.concatenate([pos_items, np.array(neg_items, dtype=np.int64)])
         self.labels = np.concatenate([pos_labels, neg_labels])
 
     def __len__(self) -> int:
